+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/handler/grid_power/ClientGridPower.kt
@@ -2,8 +2,8 @@ package dev.aaronhowser.mods.excessive_utilities.handler.grid_power
 
 object ClientGridPower {
 
-	var capacity: Int = 0
-	var usage: Int = 0
+	var capacity: Double = 0.0
+	var usage: Double = 0.0
 	val isOverloaded: Boolean get() = usage > capacity
 
 }
