+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/handler/grid_power/GridPowerContribution.kt
@@ -1,6 +1,6 @@
 package dev.aaronhowser.mods.excessive_utilities.handler.grid_power
 
 interface GridPowerContribution {
-	fun getAmount(): Int
+	fun getAmount(): Double
 	fun isStillValid(): Boolean
 }
